Fail noisily on GitHub token failures
This is a hacky fix for reviewdog/reviewdog#1210 and a partial revert of
reviewdog/reviewdog#411.

The logic for determining if a GitHub token is read-only is flawed.
Specifically it is looking to see if the PR is coming from a fork,
because in the past those were never supported.  That is no longer true,
as GitHub has added a setting to allow passing secrets to forks (which
you would do if all forks are trusted, like in a fork-based workflow
inside a company).

The idea itself is problematic because it's always going to require
keeping up to date with whatever logic GitHub is using on their backend,
so a "try it and see if it fails" approach would be more
forward-looking.

But beyond that, the entire idea of failing gracefully is an issue.  If
everything is configured correctly, great.  But if someone has
misconfigured the token, the only way they'd find out is if they go
digging into the logs on the Action; otherwise it just looks like it's
operating correctly and there are no linter warnings.  This is a problem
because it means we need to create additional monitoring on reviewdog to
see if it's working correctly, rather than just relying on the PR check
status and getting user reports when the job is failing.

An actual fix here would also strip out
`cienv.HasReadOnlyPermissionGitHubToken` and the library code that was
pulled in to handle this.  I don't know Go very well and so I've opted
to make the smallest hack I can to get this working for our situation
while waiting on the discussion about how to handle this upstream.