Install our own version of reviewdog
The installer script downloads releases from Github, rather than using
the repo that the script comes from.  This makes some amount of sense,
but is confusing in a forked-repo setup.  I'm not going to unravel all
of that to make it nicely generic, but rather just hack it for this fork
because that's a lot easier.